pyproject.toml
[tool.poetry]
name = "python-demo"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
sqlalchemy = { extras = ["asyncio", "asyncpg"], version = "^2.0.23" }
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
httpx = "^0.25.1"
itsdangerous = "^2.1.2"
asyncpg = "^0.29.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.4"
pyright = "^1.1.335"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
junit_suite_name = "pyhon-svc-poc"

[tool.pyright]
exclude = [".venv/**/*.py"]
